% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.meta.gather.task.views.R
\name{cran.meta.gather.task.views}
\alias{cran.meta.gather.task.views}
\title{Gathers the CRAN task views into a data.frame}
\usage{
cran.meta.gather.task.views()
}
\value{
A data.frame
}
\description{
Gathers the CRAN task views into a data.frame
}
