% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.meta.generate.reports.R
\name{cran.meta.generate.reports}
\alias{cran.meta.generate.reports}
\title{Generates various reports consering the CRAN meta data}
\usage{
cran.meta.generate.reports(reports.directory = NULL,
  include.code.review = FALSE)
}
\arguments{
\item{include.code.review}{boolean, include code review will install all packages to this machine this can take a long time of first run}
}
\value{
a data.frame containing CRAN metadata, and code review information
}
\description{
Generates various reports consering the CRAN meta data
}
