% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install.all.packages.from.CRAN.R
\name{cran.meta.install.all.packages}
\alias{cran.meta.install.all.packages}
\title{Installs all available packages from CRAN}
\usage{
cran.meta.install.all.packages()
}
\description{
Installs all available packages from CRAN
}
\section{as of 6/26/2019 installing all on Windows, is ~21GB, 1/2 million files, and 129k folders, but setting up a mirror of CRAN is ~210GB, this is because it holds packages for all versions}{
NA
}

