% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.meta.report.graph.R
\name{cran.meta.dependancies.graph}
\alias{cran.meta.dependancies.graph}
\title{A wraper on miniCRAN's makeDepGraph to automate generating reports, and visualizations of these graphs}
\usage{

  cran.meta.dependancies.graph(cran.packs = as.list(installed.packages()[1]),
  generate.visual.graph = TRUE,
  visual.graph.filepath = "cran.meta.report.graph.svg")
}
\arguments{
\item{cran.packs}{list containing packages to build dependancies}

\item{generate.visual.graph}{boolean determines whether or not to generate a visual graph.}

\item{visual.graph.filepath}{graph file location, will be ignored if generate.visual.graph is FALSE}
}
\value{
an igraph of the dependancy graph for supplied package list.
}
\description{
A wraper on miniCRAN's makeDepGraph to automate generating reports, and visualizations of these graphs
}
