% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.meta.install.all.packages.R
\name{cran.meta.install.all.packages}
\alias{cran.meta.install.all.packages}
\title{Install all available packages (with source) from CRAN
On first run this will take a long time (several hours), only installs packages not yet installed on your machine.
The goal is to allow us to perform in depth code reveiw of all packages on the CRAN Repository.}
\usage{
cran.meta.install.all.packages()
}
\description{
Install all available packages (with source) from CRAN
On first run this will take a long time (several hours), only installs packages not yet installed on your machine.
The goal is to allow us to perform in depth code reveiw of all packages on the CRAN Repository.
}
\note{
As of 6/26/2019 installing all on Windows, is ~19GB, 1/2 million files, and 129k folders, but setting up a mirror of CRAN is ~210GB, this is because it holds packages for all versions
}
